cleanup, added env var checks

+++ 04a_configure_aws_secrets.sh
@@ -2,28 +2,33 @@
 
 . 99_source_temp_data.sh
 
+if [ -z "$AWS_SECRET_ACCESS_KEY" ]
+then
+    echo "\$AWS_SECRET_ACCESS_KEY must be set." >&2
+	exit -1
+fi
+
+if [ -z "$AWS_ACCESS_KEY_ID" ]
+then
+    echo "\$AWS_ACCESS_KEY_ID must be set." >&2
+	exit -1
+fi
+
 tf() {
 	cd 04a_configure_aws_secrets
 	terraform init
 	terraform apply --auto-approve
 }
 
-cmd() {
-	vault read aws/config/root 
-}
-
 case $1 in
 tf)
 	tf
-	;;
-cmd)
-	cmd
 	;;
 cleanup)
 	cd 04a_configure_aws_secrets
 	terraform destroy
 	;;
 *)
-	echo "cmd for command line, tf for terraform"
+	echo "cleanup for destroy, tf for terraform"
 	;;
 esac
